@@ -16,10 +16,11 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 # In Windows, build-name is used as the major, minor, and patch parts
 # of the product and file versions while build-number is used as the build suffix.
-version: 1.0.0+1
+version: 1.0.8+8
 
 environment:
   sdk: ^3.8.1
+
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -39,7 +40,7 @@ dependencies:
   # Agregadas dependencias necesarias para tu proyecto
   provider: ^6.0.0
   flutter_date_pickers: ^0.4.3
-  supabase_flutter: ^2.5.2
+  supabase_flutter: ^2.9.1
   pdf: ^3.10.8
   printing: ^5.12.0
   share_plus: ^11.0.0
@@ -56,6 +57,10 @@ dependencies:
   flutter_contacts: ^1.1.7
   permission_handler: ^11.3.0
   collection: ^1.18.0
+  app_links: ^6.4.1
+  azlistview: ^2.0.0
+  package_info_plus: ^9.0.0
+  google_mobile_ads: ^4.0.0
 
 dev_dependencies:
   flutter_test:
@@ -80,9 +85,9 @@ dev_dependencies:
 
 flutter:
   uses-material-design: true
-  # assets:
-  #   - assets/colors_extended.json
-  #   - assets/google_logo.png
+  assets:
+    - assets/app_icon.png
+  # (Agrega aquí otros assets si los necesitas más adelante)
 
   # To add assets to your application, add an assets section, like this:
   # assets:
