@@ -1,9 +1,12 @@
 import 'dart:io';
+import 'package:app_links/app_links.dart';
+import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 import 'package:hive_flutter/hive_flutter.dart';
 import 'models/client_hive.dart';
+import 'models/contact_hive.dart';
 import 'models/transaction_hive.dart';
 
 import 'providers/client_provider.dart';
@@ -22,8 +25,14 @@ import 'providers/transaction_filter_provider.dart';
 import 'providers/tab_provider.dart';
 import 'providers/theme_provider.dart';
 import 'widgets/budgeto_theme.dart';
+import 'services/session_authority_service.dart';
+import 'services/ad_service.dart';
 
 final GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
+String _initialRoute = '/'; // Se ajustará en _initializeApp según sesión
+
+// Indicador interno para saber si ya intentamos restaurar sesión antes del árbol
+bool _preBootRestored = false;
 
 // ...eliminada duplicidad de _initializeApp...
 Future<void> _initializeApp() async {
@@ -31,11 +40,18 @@ Future<void> _initializeApp() async {
     debugPrint('Iniciando Supabase...');
     await Supabase.initialize(url: supabaseUrl, anonKey: supabaseAnonKey);
     debugPrint('Supabase inicializado');
+    // Inicia listener de depuración para cambios de estado de autenticación
+    try {
+      SessionAuthorityService.instance.startAuthStateDebugListener();
+      debugPrint('[BOOT] AuthState debug listener iniciado');
+    } catch (e) {
+      debugPrint('[BOOT] No se pudo iniciar AuthState debug listener: $e');
+    }
 
     // Captura errores globales de Supabase y de Flutter
     FlutterError.onError = (FlutterErrorDetails details) {
       if (details.exception.toString().contains('SocketException')) {
-        debugPrint('Supabase offline: \n[${details.exception}]');
+        debugPrint('Supabase offline: \n[${details.exception}]');
       } else {
         FlutterError.presentError(details);
       }
@@ -46,14 +62,64 @@ Future<void> _initializeApp() async {
     debugPrint('Hive inicializado');
     Hive.registerAdapter(ClientHiveAdapter());
     Hive.registerAdapter(TransactionHiveAdapter());
+    Hive.registerAdapter(ContactHiveAdapter());
     debugPrint('Adapters registrados');
     await Hive.openBox<ClientHive>('clients');
     debugPrint('Box clients abierto');
     await Hive.openBox<TransactionHive>('transactions');
     debugPrint('Box transactions abierto');
+    await Hive.openBox<ContactHive>('contacts');
+    debugPrint('Box contacts abierto');
     await Hive.openBox('user_settings');
+
+    // Chequeo de conectividad rápido
+    bool hasInternet = true;
+    try {
+      final r = await InternetAddress.lookup(
+        'google.com',
+      ).timeout(const Duration(seconds: 2));
+      hasInternet = r.isNotEmpty && r[0].rawAddress.isNotEmpty;
+    } catch (_) {
+      hasInternet = false;
+    }
+    // Restaurar sesión (si existe) antes de montar la UI
+    try {
+      await SessionAuthorityService.instance
+          .restoreSessionIfNeeded(hasInternet: hasInternet)
+          .timeout(const Duration(seconds: 3));
+      _preBootRestored = true;
+      debugPrint('[BOOT] Restauración previa completada');
+    } catch (e) {
+      debugPrint('[BOOT] Restauración previa falló: $e');
+    }
+
+    // Determinar initialRoute dinámicamente
+    try {
+      final session = Supabase.instance.client.auth.currentSession;
+      final clientsBox = Hive.box<ClientHive>('clients');
+      final txBox = Hive.box<TransactionHive>('transactions');
+      final hasLocalData = clientsBox.isNotEmpty || txBox.isNotEmpty;
+      if (session != null) {
+        _initialRoute = '/dashboard';
+        debugPrint('[BOOT] initialRoute => /dashboard (sesión válida)');
+      } else if (!hasInternet && hasLocalData) {
+        // Modo offline permitido directo
+        _initialRoute = '/dashboard';
+        debugPrint(
+          '[BOOT] initialRoute => /dashboard (offline con datos locales)',
+        );
+      } else {
+        _initialRoute = '/';
+        debugPrint('[BOOT] initialRoute => / (sin sesión)');
+      }
+    } catch (e) {
+      debugPrint('[BOOT] No se pudo determinar initialRoute dinamica: $e');
+    }
+
     runApp(const MyApp());
     debugPrint('runApp ejecutado');
+    // Inicializar Mobile Ads en background (no bloquear arranque)
+    AdService.instance.initialize();
   } catch (e, st) {
     runApp(
       MaterialApp(
@@ -74,8 +140,74 @@ void main() async {
   await _initializeApp();
 }
 
-class MyApp extends StatelessWidget {
+class MyApp extends StatefulWidget {
   const MyApp({super.key});
+
+  @override
+  State<MyApp> createState() => _MyAppState();
+}
+
+class _MyAppState extends State<MyApp> {
+  StreamSubscription? _sub;
+  AppLinks? _appLinks;
+  // Splash overlay
+  double _splashOpacity = 1.0;
+  bool _splashRemoved = false;
+
+  @override
+  void initState() {
+    super.initState();
+    _handleIncomingLinks();
+    // Programa el fade del overlay splash después del primer frame
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      // Pequeña espera para asegurar que initialRoute ya montó su primer frame
+      Future.delayed(const Duration(milliseconds: 120), () {
+        if (!mounted) return;
+        setState(() => _splashOpacity = 0.0);
+        // Tras el fade, remover del árbol para no costar en composición
+        Future.delayed(const Duration(milliseconds: 380), () {
+          if (mounted) setState(() => _splashRemoved = true);
+        });
+      });
+    });
+  }
+
+  void _handleIncomingLinks() {
+    _appLinks = AppLinks();
+    // Handle initial link (cold start)
+    _appLinks!
+        .getInitialLink()
+        .then((uri) {
+          if (uri != null &&
+              uri.scheme == 'deudaflow' &&
+              uri.host == 'reset-password') {
+            navigatorKey.currentState?.pushNamed('/reset-password');
+          }
+        })
+        .catchError((err) {
+          debugPrint('Deep link initial error: $err');
+          return null;
+        });
+
+    // Listen for subsequent links
+    _sub = _appLinks!.uriLinkStream.listen(
+      (Uri uri) {
+        if (uri.scheme == 'deudaflow' && uri.host == 'reset-password') {
+          navigatorKey.currentState?.pushNamed('/reset-password');
+        }
+      },
+      onError: (err) {
+        debugPrint('Deep link error: $err');
+      },
+    );
+  }
+
+  @override
+  void dispose() {
+    _sub?.cancel();
+    _appLinks = null;
+    super.dispose();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -93,8 +225,8 @@ class MyApp extends StatelessWidget {
         builder: (context, themeProvider, _) => MaterialApp(
           navigatorKey: navigatorKey,
           title: 'Deuda Flow Control',
-          theme: BudgetoTheme.light, // Usa tu tema estático
-          initialRoute: '/',
+          theme: BudgetoTheme.light,
+          initialRoute: _initialRoute,
           routes: {
             '/': (context) => const WelcomeScreen(),
             '/login': (context) => const LoginScreen(),
@@ -104,6 +236,61 @@ class MyApp extends StatelessWidget {
             '/dashboard': (context) => const AuthGate(),
             '/clients': (context) => const AuthGate(),
             '/transactions': (context) => const AuthGate(),
+          },
+          builder: (context, child) {
+            return Stack(
+              children: [
+                if (child != null) child,
+                if (!_splashRemoved)
+                  IgnorePointer(
+                    ignoring: _splashOpacity == 0.0,
+                    child: AnimatedOpacity(
+                      opacity: _splashOpacity,
+                      duration: const Duration(milliseconds: 380),
+                      curve: Curves.easeOutCubic,
+                      child: Container(
+                        decoration: const BoxDecoration(
+                          gradient: LinearGradient(
+                            begin: Alignment.topLeft,
+                            end: Alignment.bottomRight,
+                            colors: [
+                              Color(0xFF7C3AED),
+                              Color(0xFF4F46E5),
+                              Color(0xFF60A5FA),
+                            ],
+                          ),
+                        ),
+                        child: Center(
+                          child: Column(
+                            mainAxisSize: MainAxisSize.min,
+                            children: const [
+                              // Logo con fondo circular y sombra para asegurar contraste
+                              SizedBox(height: 8),
+                              Icon(
+                                Icons.account_balance_wallet_rounded,
+                                size: 130,
+                                color: Colors.white,
+                              ),
+                              SizedBox(height: 34),
+                              Text(
+                                'Deuda Flow',
+                                style: TextStyle(
+                                  fontSize: 32,
+                                  fontWeight: FontWeight.bold,
+                                  color: Colors.white,
+                                  letterSpacing: 0.8,
+                                  decoration: TextDecoration.none,
+                                  fontFamily: 'Roboto',
+                                ),
+                              ),
+                            ],
+                          ),
+                        ),
+                      ),
+                    ),
+                  ),
+              ],
+            );
           },
         ),
       ),
@@ -118,15 +305,28 @@ class AuthGate extends StatefulWidget {
 }
 
 class _AuthGateState extends State<AuthGate> {
-  Future<bool> checkInternet() async {
+  Future<Map<String, dynamic>> _bootstrap() async {
+    // Conectividad
+    bool isOnline;
     try {
       final result = await InternetAddress.lookup(
         'google.com',
       ).timeout(const Duration(seconds: 2));
-      return result.isNotEmpty && result[0].rawAddress.isNotEmpty;
+      isOnline = result.isNotEmpty && result[0].rawAddress.isNotEmpty;
     } catch (_) {
-      return false;
+      isOnline = false;
     }
+    // Si no restauramos antes (hot reload / navegaciones), intentar ahora una sola vez
+    if (!_preBootRestored) {
+      try {
+        await SessionAuthorityService.instance
+            .restoreSessionIfNeeded(hasInternet: isOnline)
+            .timeout(const Duration(seconds: 3));
+        _preBootRestored = true;
+      } catch (_) {}
+    }
+    final session = Supabase.instance.client.auth.currentSession;
+    return {'online': isOnline, 'session': session};
   }
 
   @override
@@ -135,19 +335,19 @@ class _AuthGateState extends State<AuthGate> {
     final transactionsBox = Hive.box<TransactionHive>('transactions');
     final hasLocalData = clientsBox.isNotEmpty || transactionsBox.isNotEmpty;
 
-    return FutureBuilder<bool>(
-      future: checkInternet(),
+    return FutureBuilder<Map<String, dynamic>>(
+      future: _bootstrap(),
       builder: (context, snapshot) {
         if (!snapshot.hasData) {
           return const Center(child: CircularProgressIndicator());
         }
-        final isOnline = snapshot.data ?? false;
+        final data = snapshot.data!;
+        final isOnline = data['online'] as bool;
+        final session = data['session'];
         if (!isOnline && hasLocalData) {
           // Modo offline: acceso directo
           return MainScaffold(userId: 'offline');
         }
-        // Si hay internet, consulta Supabase normalmente
-        final session = Supabase.instance.client.auth.currentSession;
         if (session == null) {
           return const LoginScreen();
         } else {
