@@ -25,14 +25,15 @@ class ClientHiveAdapter extends TypeAdapter<ClientHive> {
       synced: fields[5] as bool,
       pendingDelete: fields[6] as bool,
       localId: fields[7] as String?,
-      currencyCode: fields[8] as String,
+      currencyCode: fields[8] as String?,
+      anchorUsdValue: fields[9] as double?,
     );
   }
 
   @override
   void write(BinaryWriter writer, ClientHive obj) {
     writer
-      ..writeByte(9)
+      ..writeByte(10)
       ..writeByte(0)
       ..write(obj.id)
       ..writeByte(1)
@@ -50,7 +51,9 @@ class ClientHiveAdapter extends TypeAdapter<ClientHive> {
       ..writeByte(7)
       ..write(obj.localId)
       ..writeByte(8)
-      ..write(obj.currencyCode);
+      ..write(obj.currencyCode)
+      ..writeByte(9)
+      ..write(obj.anchorUsdValue);
   }
 
   @override
