@@ -2,6 +2,8 @@ import 'package:shared_preferences/shared_preferences.dart';
 import '../widgets/faq_help_sheet.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart'; // Importar para formateo de números
+import 'package:characters/characters.dart';
+import '../utils/string_sanitizer.dart';
 import 'package:provider/provider.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 import '../models/client.dart';
@@ -14,6 +16,7 @@ import '../widgets/cyclic_animated_fade_list.dart';
 import '../widgets/dashboard_stats.dart';
 import '../widgets/sync_banner.dart';
 import '../services/supabase_service.dart';
+import '../services/session_authority_service.dart';
 
 // Para kIsWeb
 class DashboardScreen extends StatefulWidget {
@@ -31,8 +34,17 @@ class DashboardScreen extends StatefulWidget {
   State<DashboardScreen> createState() => _DashboardScreenState();
 }
 
-class _DashboardScreenState extends State<DashboardScreen> {
+class _DashboardScreenState extends State<DashboardScreen>
+    with SingleTickerProviderStateMixin {
   bool _faqShown = false;
+  String _capitalizeSafe(String input) {
+    final chars = input.characters;
+    if (chars.isEmpty) return '';
+    final first = chars.first.toUpperCase();
+    final rest = chars.skip(1).toString().toLowerCase();
+    return '$first$rest';
+  }
+
   void _showRateDialog(BuildContext context, double initialRate) {
     final TextEditingController rateController = TextEditingController(
       text: initialRate.toString(),
@@ -82,10 +94,29 @@ class _DashboardScreenState extends State<DashboardScreen> {
   }
 
   bool _loading = true;
+  late final AnimationController _syncController;
+  void _syncStatusListener() {
+    try {
+      final sp = Provider.of<SyncProvider>(context, listen: false);
+      if (sp.status.toString().contains('syncing')) {
+        if (!_syncController.isAnimating) _syncController.repeat();
+      } else {
+        if (_syncController.isAnimating) {
+          _syncController.reset();
+        }
+      }
+    } catch (_) {
+      // Context might be unavailable during dispose
+    }
+  }
 
   @override
   void initState() {
     super.initState();
+    _syncController = AnimationController(
+      vsync: this,
+      duration: const Duration(seconds: 1),
+    );
     _loadData();
     // Inicializa el estado de SyncProvider para que el banner se muestre correctamente
     Future.microtask(() {
@@ -98,6 +129,31 @@ class _DashboardScreenState extends State<DashboardScreen> {
     // Comprobar diferencias de balance tras cargar datos
     Future.delayed(const Duration(milliseconds: 800), _checkBalanceDifferences);
     // El FAQ solo se mostrará después de cargar datos y si el usuario está autenticado
+
+    // --- INICIO LISTENER DE DEVICE_ID ---
+    final user = Supabase.instance.client.auth.currentUser;
+    if (user != null) {
+      // Se recomienda usar addPostFrameCallback para asegurar que el context es válido
+      WidgetsBinding.instance.addPostFrameCallback((_) {
+        SessionAuthorityService.instance.listenToDeviceIdChanges(
+          user.id,
+          context,
+        );
+        // Añadir listener de SyncProvider para animar el icono de sync
+        final sp = Provider.of<SyncProvider>(context, listen: false);
+        sp.addListener(_syncStatusListener);
+      });
+    }
+  }
+
+  @override
+  void dispose() {
+    try {
+      final sp = Provider.of<SyncProvider>(context, listen: false);
+      sp.removeListener(_syncStatusListener);
+    } catch (_) {}
+    _syncController.dispose();
+    super.dispose();
   }
 
   Future<void> _showFaqIfFirstTime() async {
@@ -115,59 +171,199 @@ class _DashboardScreenState extends State<DashboardScreen> {
     }
   }
 
-  // Compara balances locales y remotos y muestra SnackBar si hay diferencias
+  // Compara balances locales y remotos y muestra un diálogo si hay diferencias
   Future<void> _checkBalanceDifferences() async {
-    // El siguiente uso de context es seguro porque:
-    // 1. Se verifica 'if (difference && mounted)' antes de usar context tras el async gap.
-    // 2. Este context es el de la clase State, no de un builder externo.
-    // Por lo tanto, el warning puede ser ignorado.
-    // ignore: use_build_context_synchronously
-    final clientProvider = Provider.of<ClientProvider>(
-      // El siguiente uso de context es seguro porque:
-      // 1. Se verifica 'if (difference && mounted)' antes de usar context tras el async gap.
-      // 2. Este context es el de la clase State, no de un builder externo.
-      // Por lo tanto, el warning puede ser ignorado.
-      // ignore: use_build_context_synchronously
-      context,
-      listen: false,
-    );
+    final clientProvider = Provider.of<ClientProvider>(context, listen: false);
     final localClients = clientProvider.clients;
     try {
-      // Asume que tienes un método en SupabaseService para obtener clientes remotos
       final supabaseService = SupabaseService();
       final remoteClients = await supabaseService.fetchClients(widget.userId);
-      // Mapear por id para comparar
       final localMap = {for (var c in localClients) c.id: c};
       final remoteMap = {for (var c in remoteClients) c.id: c};
-      bool difference = false;
+      // Detectar clientes con conflicto
+      final List<Map<String, dynamic>> conflicts = [];
       for (final id in localMap.keys) {
         if (remoteMap.containsKey(id)) {
           final localBal = localMap[id]!.balance;
           final remoteBal = remoteMap[id]!.balance;
+          debugPrint(
+            '[BALANCE] Cliente: ${localMap[id]!.name} (id: $id) | Local: $localBal | Supabase: $remoteBal',
+          );
           if ((localBal - remoteBal).abs() > 0.01) {
-            difference = true;
-            break;
+            debugPrint(
+              '[CONFLICTO BALANCE] Cliente: ${localMap[id]!.name} (id: $id) | Local: $localBal | Supabase: $remoteBal',
+            );
+            conflicts.add({
+              'id': id,
+              'name': localMap[id]!.name,
+              'local': localBal,
+              'remote': remoteBal,
+            });
           }
         }
       }
-      if (difference && mounted) {
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(
-            content: const Text(
-              '¡Atención! Hay diferencias entre los balances locales y online.',
-              style: TextStyle(fontWeight: FontWeight.bold),
-            ),
-            duration: const Duration(minutes: 5),
-            action: SnackBarAction(
-              label: 'X',
-              onPressed: () {
-                ScaffoldMessenger.of(context).hideCurrentSnackBar();
-              },
-              textColor: Colors.white,
-            ),
-            backgroundColor: Colors.deepOrange,
-            behavior: SnackBarBehavior.floating,
-          ),
+      if (conflicts.isNotEmpty && mounted) {
+        await showDialog(
+          context: context,
+          barrierDismissible: false,
+          builder: (ctx) {
+            return AlertDialog(
+              shape: RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(16),
+              ),
+              title: const Text(
+                'Conflicto de balances',
+                style: TextStyle(fontWeight: FontWeight.bold),
+              ),
+              content: Builder(
+                builder: (ctx) {
+                  final maxHeight = MediaQuery.of(ctx).size.height * 0.6;
+                  return SingleChildScrollView(
+                    child: ConstrainedBox(
+                      constraints: BoxConstraints(
+                        maxWidth: 340,
+                        maxHeight: maxHeight,
+                      ),
+                      child: Column(
+                        mainAxisSize: MainAxisSize.min,
+                        crossAxisAlignment: CrossAxisAlignment.start,
+                        children: [
+                          const Text(
+                            'Se detectaron diferencias entre los balances locales y online de los siguientes clientes:',
+                            style: TextStyle(fontSize: 15),
+                          ),
+                          const SizedBox(height: 12),
+                          // La lista se envuelve en Flexible para respetar el maxHeight y poder hacer scroll
+                          Flexible(
+                            child: ListView.builder(
+                              shrinkWrap: true,
+                              physics: const AlwaysScrollableScrollPhysics(),
+                              itemCount: conflicts.length,
+                              itemBuilder: (context, i) {
+                                final c = conflicts[i];
+                                return Padding(
+                                  padding: const EdgeInsets.symmetric(
+                                    vertical: 4.0,
+                                  ),
+                                  child: Row(
+                                    children: [
+                                      Expanded(
+                                        child: Text(
+                                          c['name'] ?? '',
+                                          style: const TextStyle(
+                                            fontWeight: FontWeight.bold,
+                                          ),
+                                        ),
+                                      ),
+                                      Column(
+                                        crossAxisAlignment:
+                                            CrossAxisAlignment.end,
+                                        children: [
+                                          Text(
+                                            'Local: ${c['local'].toStringAsFixed(2)}',
+                                            style: const TextStyle(
+                                              color: Colors.blue,
+                                            ),
+                                          ),
+                                          Text(
+                                            'Online: ${c['remote'].toStringAsFixed(2)}',
+                                            style: const TextStyle(
+                                              color: Colors.deepOrange,
+                                            ),
+                                          ),
+                                        ],
+                                      ),
+                                    ],
+                                  ),
+                                );
+                              },
+                            ),
+                          ),
+                          const SizedBox(height: 10),
+                          const Text(
+                            '¿Con qué versión deseas quedarte para todos los clientes?',
+                            style: TextStyle(fontWeight: FontWeight.w500),
+                          ),
+                        ],
+                      ),
+                    ),
+                  );
+                },
+              ),
+              actionsPadding: const EdgeInsets.symmetric(vertical: 8),
+              actions: [
+                Center(
+                  child: Consumer<SyncProvider>(
+                    builder: (context, syncProvider, _) {
+                      final isSyncing = syncProvider.status.toString().contains(
+                        'syncing',
+                      );
+                      final onlineFuture = Provider.of<TransactionProvider>(
+                        context,
+                        listen: false,
+                      ).isOnline();
+                      return FutureBuilder<bool>(
+                        future: onlineFuture,
+                        builder: (context, snapshot) {
+                          final online = snapshot.data ?? true;
+                          if (!online) return const SizedBox.shrink();
+                          return ElevatedButton(
+                            onPressed: isSyncing
+                                ? () {}
+                                : () {
+                                    // Llamar al SyncProvider para forzar sincronización completa
+                                    try {
+                                      Provider.of<SyncProvider>(
+                                        context,
+                                        listen: false,
+                                      ).startSync(context, widget.userId);
+                                    } catch (_) {}
+                                    if (mounted) Navigator.of(ctx).pop();
+                                  },
+                            style: ElevatedButton.styleFrom(
+                              backgroundColor: Theme.of(
+                                context,
+                              ).colorScheme.primary,
+                              shape: RoundedRectangleBorder(
+                                borderRadius: BorderRadius.circular(10),
+                              ),
+                              padding: const EdgeInsets.symmetric(
+                                horizontal: 18,
+                                vertical: 12,
+                              ),
+                            ),
+                            child: AnimatedBuilder(
+                              animation: _syncController,
+                              builder: (context, child) {
+                                return Row(
+                                  mainAxisSize: MainAxisSize.min,
+                                  children: [
+                                    Transform.rotate(
+                                      angle: isSyncing
+                                          ? -_syncController.value * 6.28319
+                                          : 0,
+                                      child: Icon(
+                                        Icons.sync,
+                                        color: isSyncing
+                                            ? Colors.green
+                                            : Colors.white,
+                                      ),
+                                    ),
+                                    const SizedBox(width: 10),
+                                    const Text('Sincronizar Todo'),
+                                  ],
+                                );
+                              },
+                            ),
+                          );
+                        },
+                      );
+                    },
+                  ),
+                ),
+              ],
+            );
+          },
         );
       }
     } catch (e) {
@@ -207,23 +403,16 @@ class _DashboardScreenState extends State<DashboardScreen> {
           (meta['name'] as String).trim().isNotEmpty) {
         userName = (meta['name'] as String)
             .trim()
-            .split(' ')
-            .map(
-              (w) => w.isNotEmpty
-                  ? w[0].toUpperCase() + w.substring(1).toLowerCase()
-                  : '',
-            )
+            .split(RegExp(r'\s+'))
+            .map((w) => w.isNotEmpty ? _capitalizeSafe(w) : '')
             .join(' ');
       } else if (user.email != null) {
         final emailName = user.email!.split('@')[0];
-        userName = emailName.isNotEmpty
-            ? emailName[0].toUpperCase() + emailName.substring(1)
-            : '';
+        userName = emailName.isNotEmpty ? _capitalizeSafe(emailName) : '';
       }
     } else {
       userName = 'Invitado';
     }
-
     String saludo() {
       final hour = DateTime.now().hour;
       if (hour >= 5 && hour < 12) return 'Buenos días';
@@ -397,35 +586,76 @@ class _DashboardScreenState extends State<DashboardScreen> {
                                             CrossAxisAlignment.start,
                                         children: [
                                           const SizedBox(height: 8),
-                                          Text(
-                                            userName.isNotEmpty
-                                                ? '${saludo()}, $userName'
-                                                : saludo(),
-                                            style: const TextStyle(
-                                              color: Color.fromARGB(
-                                                255,
-                                                255,
-                                                255,
-                                                255,
-                                              ),
-                                              fontSize: 32,
-                                              fontWeight: FontWeight.bold,
-                                              letterSpacing: 0.5,
-                                              shadows: [
-                                                Shadow(
-                                                  color: Color.fromARGB(
-                                                    0,
-                                                    247,
-                                                    246,
-                                                    246,
+                                          FittedBox(
+                                            fit: BoxFit.scaleDown,
+                                            alignment: Alignment.centerLeft,
+                                            child: Column(
+                                              crossAxisAlignment:
+                                                  CrossAxisAlignment.start,
+                                              mainAxisSize: MainAxisSize.min,
+                                              children: [
+                                                Text(
+                                                  saludo(),
+                                                  style: const TextStyle(
+                                                    color: Color.fromARGB(
+                                                      255,
+                                                      255,
+                                                      255,
+                                                      255,
+                                                    ),
+                                                    fontSize: 32,
+                                                    fontWeight: FontWeight.bold,
+                                                    letterSpacing: 0.5,
+                                                    shadows: [
+                                                      Shadow(
+                                                        color: Color.fromARGB(
+                                                          0,
+                                                          247,
+                                                          246,
+                                                          246,
+                                                        ),
+                                                        offset: Offset(0, 0),
+                                                        blurRadius: 8,
+                                                      ),
+                                                    ],
                                                   ),
-                                                  offset: Offset(0, 0),
-                                                  blurRadius: 8,
+                                                  maxLines: 1,
+                                                  overflow:
+                                                      TextOverflow.ellipsis,
                                                 ),
+                                                if (userName.isNotEmpty)
+                                                  Text(
+                                                    userName,
+                                                    style: const TextStyle(
+                                                      color: Color.fromARGB(
+                                                        255,
+                                                        255,
+                                                        255,
+                                                        255,
+                                                      ),
+                                                      fontSize: 32,
+                                                      fontWeight:
+                                                          FontWeight.bold,
+                                                      letterSpacing: 0.5,
+                                                      shadows: [
+                                                        Shadow(
+                                                          color: Color.fromARGB(
+                                                            0,
+                                                            247,
+                                                            246,
+                                                            246,
+                                                          ),
+                                                          offset: Offset(0, 0),
+                                                          blurRadius: 8,
+                                                        ),
+                                                      ],
+                                                    ),
+                                                    maxLines: 1,
+                                                    overflow:
+                                                        TextOverflow.ellipsis,
+                                                  ),
                                               ],
                                             ),
-                                            maxLines: 2,
-                                            overflow: TextOverflow.ellipsis,
                                           ),
                                           const SizedBox(height: 12),
                                         ],
@@ -527,7 +757,9 @@ class _DashboardScreenState extends State<DashboardScreen> {
                                         ),
                                       );
                                       final clientName = client.name.isNotEmpty
-                                          ? client.name
+                                          ? StringSanitizer.sanitizeForText(
+                                              client.name,
+                                            )
                                           : 'Desconocido';
 
                                       // --- Lógica de balance corregida ---
