@@ -26,7 +26,6 @@ class _LoginScreenState extends State<LoginScreen> {
   // Animación simple para los botones: escala al presionar
   double _loginBtnScale = 1.0;
   double _offlineBtnScale = 1.0;
-  double _guestBtnScale = 1.0;
   double _googleBtnScale = 1.0;
 
   @override
@@ -77,31 +76,6 @@ class _LoginScreenState extends State<LoginScreen> {
   }
 
   // Datos de invitado (ajusta si es necesario)
-  static const String _guestEmail = 'invitado@deudaflow.com';
-  static const String _guestPassword = 'invitado123';
-
-  Future<void> _loginAsGuest() async {
-    _emailController.text = _guestEmail;
-    _passwordController.text = _guestPassword;
-    // Verifica si ya hay sesión guardada para invitado
-    final sessionBox = await Hive.openBox('session');
-    final savedEmail = sessionBox.get('email');
-    if (savedEmail == _guestEmail) {
-      // Ya hay sesión guardada, permite acceso offline
-      await _login();
-    } else {
-      // No hay sesión guardada, requiere internet la primera vez
-      if (!mounted) return;
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(
-          content: Text(
-            'No es posible usar el modo invitado sin conexión la primera vez. Por favor, conéctate a internet e inicia sesión como invitado para habilitar el acceso offline.',
-          ),
-          duration: Duration(seconds: 4),
-        ),
-      );
-    }
-  }
 
   final _emailController = TextEditingController();
   final _passwordController = TextEditingController();
@@ -133,6 +107,14 @@ class _LoginScreenState extends State<LoginScreen> {
     if (email.isEmpty || password.isEmpty) {
       setState(() {
         _error = 'Email y contraseña requeridos';
+        _loading = false;
+      });
+      return;
+    }
+    // Verificar conectividad antes de intentar login online
+    if (!await _hasConnectivity()) {
+      setState(() {
+        _error = 'Sin conexión a internet.';
         _loading = false;
       });
       return;
@@ -217,6 +199,8 @@ class _LoginScreenState extends State<LoginScreen> {
           await txProvider.syncPendingTransactions(user.id);
         } catch (_) {}
         if (!mounted) return;
+        // Guardar sesión Supabase en Hive (JSON) para restauración futura
+        await SessionAuthorityService.instance.saveSupabaseSessionLocally();
         // ignore: use_build_context_synchronously
         Navigator.of(context).pushReplacementNamed('/dashboard');
       }
@@ -234,33 +218,18 @@ class _LoginScreenState extends State<LoginScreen> {
           Navigator.of(context).pushReplacementNamed('/dashboard');
         } else {
           // Si es invitado, mensaje especial
-          if (email == _guestEmail) {
-            setState(() {
-              _error =
-                  'No es posible usar el modo invitado sin conexión la primera vez. Conéctate a internet e inicia sesión como invitado para habilitar el acceso offline.';
-            });
-          } else {
-            setState(() {
-              _error = 'No hay sesión guardada para este usuario.';
-            });
-          }
+          setState(() {
+            _error = 'No hay sesión guardada para este usuario.';
+          });
         }
       } else {
         if (!mounted) return;
         // Si es invitado, mensaje más amigable
-        if (email == _guestEmail) {
-          setState(() {
-            _error =
-                'No se pudo acceder como invitado. Verifica tu conexión o intenta más tarde.';
-          });
-        } else {
-          setState(() {
-            _error =
-                e.message.toLowerCase().contains('invalid login credentials')
-                ? 'Credenciales incorrectas. Verifica tu email y contraseña.'
-                : e.message;
-          });
-        }
+        setState(() {
+          _error = e.message.toLowerCase().contains('invalid login credentials')
+              ? 'Credenciales incorrectas. Verifica tu email y contraseña.'
+              : e.message;
+        });
       }
     } catch (e) {
       // Si es error de red, permitir login offline si hay datos guardadas
@@ -275,29 +244,15 @@ class _LoginScreenState extends State<LoginScreen> {
           Navigator.of(context).pushReplacementNamed('/dashboard');
         } else {
           // Si es invitado, mensaje especial
-          if (email == _guestEmail) {
-            setState(() {
-              _error =
-                  'No es posible usar el modo invitado sin conexión la primera vez. Conéctate a internet e inicia sesión como invitado para habilitar el acceso offline.';
-            });
-          } else {
-            setState(() {
-              _error = 'No hay sesión guardada para este usuario.';
-            });
-          }
+          setState(() {
+            _error = 'No hay sesión guardada para este usuario.';
+          });
         }
       } else {
         // Si es invitado, mensaje más amigable
-        if (email == _guestEmail) {
-          setState(() {
-            _error =
-                'No se pudo acceder como invitado. Verifica tu conexión o intenta más tarde.';
-          });
-        } else {
-          setState(() {
-            _error = 'Error inesperado: ${e.toString()}';
-          });
-        }
+        setState(() {
+          _error = 'Error inesperado: ${e.toString()}';
+        });
       }
     } finally {
       setState(() {
@@ -376,6 +331,7 @@ class _LoginScreenState extends State<LoginScreen> {
         idToken: idToken,
       );
       debugPrint('DEBUG: Supabase signInWithIdToken completado: $res');
+      debugPrint('Supabase signIn response: ${res.session}');
       final user = res.user;
       if (user == null) {
         debugPrint('DEBUG: Supabase devolvió user == null');
@@ -431,6 +387,7 @@ class _LoginScreenState extends State<LoginScreen> {
         }
       } catch (_) {}
       debugPrint('DEBUG: Login con Google exitoso, navegando a dashboard.');
+      await SessionAuthorityService.instance.saveSupabaseSessionLocally();
       // ignore: use_build_context_synchronously
       Navigator.of(context).pushReplacementNamed('/dashboard');
     } catch (e) {
@@ -648,148 +605,8 @@ class _LoginScreenState extends State<LoginScreen> {
                             ),
                           ),
                         ),
-                        const SizedBox(height: 10),
-                        Row(
-                          mainAxisAlignment: MainAxisAlignment.center,
-                          children: [
-                            Expanded(
-                              child: TextButton(
-                                onPressed: () => Navigator.of(context).push(
-                                  MaterialPageRoute(
-                                    builder: (_) => const RegisterScreen(),
-                                  ),
-                                ),
-                                child: const Text(
-                                  'Registrarse',
-                                  overflow: TextOverflow.ellipsis,
-                                  style: TextStyle(color: Colors.white),
-                                ),
-                              ),
-                            ),
-                            Expanded(
-                              child: TextButton(
-                                onPressed: () => Navigator.of(context).push(
-                                  MaterialPageRoute(
-                                    builder: (_) =>
-                                        const ForgotPasswordScreen(),
-                                  ),
-                                ),
-                                child: const Text(
-                                  '¿Olvidaste tu contraseña?',
-                                  maxLines: 2,
-                                  overflow: TextOverflow.ellipsis,
-                                  textAlign: TextAlign.center,
-                                  style: TextStyle(color: Colors.white),
-                                ),
-                              ),
-                            ),
-                          ],
-                        ),
-                        const SizedBox(height: 8),
-                        SizedBox(
-                          width: double.infinity,
-                          child: GestureDetector(
-                            onTapDown: (_) =>
-                                setState(() => _offlineBtnScale = 0.93),
-                            onTapUp: (_) =>
-                                setState(() => _offlineBtnScale = 1.0),
-                            onTapCancel: () =>
-                                setState(() => _offlineBtnScale = 1.0),
-                            onTap: _loading
-                                ? null
-                                : () {
-                                    setState(() => _offlineBtnScale = 1.0);
-                                    _loginOffline();
-                                  },
-                            child: AnimatedScale(
-                              scale: _offlineBtnScale,
-                              duration: const Duration(milliseconds: 120),
-                              curve: Curves.easeOut,
-                              child: OutlinedButton.icon(
-                                onPressed:
-                                    null, // Desactivado, solo GestureDetector ejecuta la acción
-                                icon: const Icon(
-                                  Icons.wifi_off,
-                                  size: 18,
-                                  color: Colors.white,
-                                ),
-                                label: const Text(
-                                  'Ingresar sin conexión',
-                                  style: TextStyle(
-                                    fontSize: 14,
-                                    color: Colors.white,
-                                  ),
-                                ),
-                                style: OutlinedButton.styleFrom(
-                                  side: const BorderSide(
-                                    color: Colors.white70,
-                                    width: 1.2,
-                                  ),
-                                  minimumSize: const Size(0, 38),
-                                  padding: const EdgeInsets.symmetric(
-                                    vertical: 8,
-                                  ),
-                                  shape: RoundedRectangleBorder(
-                                    borderRadius: BorderRadius.circular(8),
-                                  ),
-                                ),
-                              ),
-                            ),
-                          ),
-                        ),
                         const SizedBox(height: 6),
-                        SizedBox(
-                          width: double.infinity,
-                          child: GestureDetector(
-                            onTapDown: (_) =>
-                                setState(() => _guestBtnScale = 0.93),
-                            onTapUp: (_) =>
-                                setState(() => _guestBtnScale = 1.0),
-                            onTapCancel: () =>
-                                setState(() => _guestBtnScale = 1.0),
-                            onTap: _loading
-                                ? null
-                                : () {
-                                    setState(() => _guestBtnScale = 1.0);
-                                    _loginAsGuest();
-                                  },
-                            child: AnimatedScale(
-                              scale: _guestBtnScale,
-                              duration: const Duration(milliseconds: 120),
-                              curve: Curves.easeOut,
-                              child: OutlinedButton.icon(
-                                onPressed:
-                                    null, // Desactivado, solo GestureDetector ejecuta la acción
-                                icon: const Icon(
-                                  Icons.person_outline,
-                                  size: 18,
-                                  color: Colors.white,
-                                ),
-                                label: const Text(
-                                  'Probar como invitado',
-                                  style: TextStyle(
-                                    fontSize: 14,
-                                    color: Colors.white,
-                                  ),
-                                ),
-                                style: OutlinedButton.styleFrom(
-                                  side: const BorderSide(
-                                    color: Colors.white70,
-                                    width: 1.2,
-                                  ),
-                                  minimumSize: const Size(0, 38),
-                                  padding: const EdgeInsets.symmetric(
-                                    vertical: 8,
-                                  ),
-                                  shape: RoundedRectangleBorder(
-                                    borderRadius: BorderRadius.circular(8),
-                                  ),
-                                ),
-                              ),
-                            ),
-                          ),
-                        ),
-                        const SizedBox(height: 6),
+                        // ...guest login button removed...
                         SizedBox(
                           width: double.infinity,
                           child: GestureDetector(
@@ -843,7 +660,93 @@ class _LoginScreenState extends State<LoginScreen> {
                           ),
                         ),
                         const SizedBox(height: 6),
-                        // ...existing code...
+                        SizedBox(
+                          width: double.infinity,
+                          child: GestureDetector(
+                            onTapDown: (_) =>
+                                setState(() => _offlineBtnScale = 0.93),
+                            onTapUp: (_) =>
+                                setState(() => _offlineBtnScale = 1.0),
+                            onTapCancel: () =>
+                                setState(() => _offlineBtnScale = 1.0),
+                            onTap: _loading
+                                ? null
+                                : () {
+                                    setState(() => _offlineBtnScale = 1.0);
+                                    _loginOffline();
+                                  },
+                            child: AnimatedScale(
+                              scale: _offlineBtnScale,
+                              duration: const Duration(milliseconds: 120),
+                              curve: Curves.easeOut,
+                              child: OutlinedButton.icon(
+                                onPressed:
+                                    null, // Desactivado, solo GestureDetector ejecuta la acción
+                                icon: const Icon(
+                                  Icons.wifi_off,
+                                  size: 18,
+                                  color: Colors.white,
+                                ),
+                                label: const Text(
+                                  'Ingresar sin conexión',
+                                  style: TextStyle(
+                                    fontSize: 14,
+                                    color: Colors.white,
+                                  ),
+                                ),
+                                style: OutlinedButton.styleFrom(
+                                  side: const BorderSide(
+                                    color: Colors.white70,
+                                    width: 1.2,
+                                  ),
+                                  minimumSize: const Size(0, 38),
+                                  padding: const EdgeInsets.symmetric(
+                                    vertical: 8,
+                                  ),
+                                  shape: RoundedRectangleBorder(
+                                    borderRadius: BorderRadius.circular(8),
+                                  ),
+                                ),
+                              ),
+                            ),
+                          ),
+                        ),
+                        Row(
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          children: [
+                            Expanded(
+                              child: TextButton(
+                                onPressed: () => Navigator.of(context).push(
+                                  MaterialPageRoute(
+                                    builder: (_) => const RegisterScreen(),
+                                  ),
+                                ),
+                                child: const Text(
+                                  'Registrarse',
+                                  overflow: TextOverflow.ellipsis,
+                                  style: TextStyle(color: Colors.white),
+                                ),
+                              ),
+                            ),
+                            Expanded(
+                              child: TextButton(
+                                onPressed: () => Navigator.of(context).push(
+                                  MaterialPageRoute(
+                                    builder: (_) =>
+                                        const ForgotPasswordScreen(),
+                                  ),
+                                ),
+                                child: const Text(
+                                  '¿Olvidaste tu contraseña?',
+                                  maxLines: 2,
+                                  overflow: TextOverflow.ellipsis,
+                                  textAlign: TextAlign.center,
+                                  style: TextStyle(color: Colors.white),
+                                ),
+                              ),
+                            ),
+                          ],
+                        ), // ...existing code...
                       ],
                     ),
                   ),
