@@ -1,3 +1,4 @@
+import '../utils/currency_utils.dart';
 import 'package:flutter/material.dart';
 import '../widgets/budgeto_colors.dart';
 import 'package:flutter/services.dart';
@@ -7,6 +8,47 @@ import '../models/transaction.dart';
 import '../providers/client_provider.dart';
 import '../providers/transaction_provider.dart';
 import '../providers/currency_provider.dart';
+
+import 'package:intl/intl.dart';
+
+/// Formatter para mostrar el monto con puntos de miles y coma decimal en tiempo real.
+/// Ejemplo: 1234567,89 -> 1.234.567,89
+class ThousandsFormatter extends TextInputFormatter {
+  @override
+  TextEditingValue formatEditUpdate(
+    TextEditingValue oldValue,
+    TextEditingValue newValue,
+  ) {
+    // Permitir borrar todo
+    if (newValue.text.isEmpty) {
+      return newValue.copyWith(text: '');
+    }
+    // Permitir solo números y una coma
+    final reg = RegExp(r'^[0-9]{1,}(,[0-9]{0,2})? 0');
+    final raw = newValue.text.replaceAll('.', '');
+    // Si el usuario está escribiendo, no formatear hasta que sea válido
+    if (!RegExp(r'^[0-9]+(,[0-9]{0,2})?$').hasMatch(raw)) {
+      return oldValue;
+    }
+    final parts = raw.split(',');
+    final intPart = parts[0];
+    String formattedInt = NumberFormat(
+      '#,###',
+      'es',
+    ).format(int.parse(intPart)).replaceAll(',', '.');
+    String formatted = formattedInt;
+    if (parts.length > 1) {
+      String decimal = parts[1];
+      if (decimal.length > 2) decimal = decimal.substring(0, 2);
+      formatted += ',' + decimal;
+    }
+    // No forzar la coma y ceros hasta que el usuario termine
+    return TextEditingValue(
+      text: formatted,
+      selection: TextSelection.collapsed(offset: formatted.length),
+    );
+  }
+}
 
 class AddGlobalTransactionModal extends StatelessWidget {
   final String userId;
@@ -53,7 +95,9 @@ class _GlobalTransactionForm extends StatefulWidget {
 
 class _GlobalTransactionFormState extends State<_GlobalTransactionForm> {
   final _amountController = TextEditingController();
+  final _amountFocusNode = FocusNode();
   final _descriptionController = TextEditingController();
+  final _thousandsFormatter = ThousandsFormatter();
   String? _type;
   String? _currencyCode;
   // FIX: Normalizar la fecha inicial a medianoche para evitar que la hora interfiera con el ordenamiento.
@@ -68,7 +112,49 @@ class _GlobalTransactionFormState extends State<_GlobalTransactionForm> {
   final _rateController = TextEditingController();
   bool _rateFieldVisible = false;
 
-  // Usar logger en vez de print para errores y advertencias
+  @override
+  void initState() {
+    super.initState();
+    _amountFocusNode.addListener(() {
+      if (!_amountFocusNode.hasFocus) {
+        // Al perder el foco, si no hay decimales, agregar ',00'
+        String raw = _amountController.text.replaceAll('.', '');
+        if (raw.isEmpty) return;
+        final parts = raw.split(',');
+        String intPart = parts[0];
+        String formattedInt = NumberFormat(
+          '#,###',
+          'es',
+        ).format(int.parse(intPart)).replaceAll(',', '.');
+        String formatted = formattedInt;
+        if (parts.length > 1) {
+          String decimal = parts[1];
+          if (decimal.length > 2) decimal = decimal.substring(0, 2);
+          formatted += ',' + decimal;
+          if (decimal.length == 0)
+            formatted += '00';
+          else if (decimal.length == 1)
+            formatted += '0';
+        } else {
+          formatted += ',00';
+        }
+        _amountController.text = formatted;
+        _amountController.selection = TextSelection.collapsed(
+          offset: formatted.length,
+        );
+      }
+    });
+  }
+
+  @override
+  void dispose() {
+    _amountFocusNode.dispose();
+    _amountController.dispose();
+    _descriptionController.dispose();
+    _rateController.dispose();
+    super.dispose();
+  }
+
   Future<void> _save() async {
     setState(() {
       _error = null;
@@ -102,8 +188,22 @@ class _GlobalTransactionFormState extends State<_GlobalTransactionForm> {
       logError('Debes seleccionar una moneda');
       return;
     }
-    final amount = double.tryParse(_amountController.text);
-    if (amount == null || amount <= 0) {
+    // Validar monto como string formateado
+    final amountStr = _amountController.text;
+    if (amountStr.isEmpty ||
+        !RegExp(r'^\d{1,3}(\.\d{3})*(,\d{1,2})?$').hasMatch(amountStr)) {
+      setState(() {
+        _error = 'Monto inválido';
+        _loading = false;
+      });
+      logError('Monto inválido');
+      return;
+    }
+    // Convertir el string formateado a double para guardar en Transaction.amount
+    final amountDouble = double.tryParse(
+      amountStr.replaceAll('.', '').replaceAll(',', '.'),
+    );
+    if (amountDouble == null || amountDouble <= 0) {
       setState(() {
         _error = 'Monto inválido';
         _loading = false;
@@ -177,24 +277,27 @@ class _GlobalTransactionFormState extends State<_GlobalTransactionForm> {
         listen: false,
       );
       double? rate;
+      double? amount = amountDouble;
       if (_currencyCode != null) {
-        final codeUC = _currencyCode!.toUpperCase();
-        rate = currencyProvider.exchangeRates[codeUC];
-        if (rate != null && rate > 0) {
-          anchorUsdValue = amount / rate;
+        // Usar el método robusto del provider para obtener la tasa
+        rate = currencyProvider.getRateFor(_currencyCode!);
+        if (_currencyCode!.toUpperCase() == 'USD') {
+          anchorUsdValue = CurrencyUtils.normalizeAnchorUsd(amount);
           debugPrint(
-            '\u001b[41m[GLOBAL_FORM][CALC] amount=$amount, currency=$_currencyCode, rate=$rate, anchorUsdValue=$anchorUsdValue\u001b[0m',
+            '\u001b[41m[GLOBAL_FORM][CALC] amount=$amountStr, currency=USD, anchorUsdValue=$anchorUsdValue\u001b[0m',
           );
-        } else if (codeUC == 'USD') {
-          anchorUsdValue = amount;
+        } else if (rate != null && rate > 0) {
+          anchorUsdValue = CurrencyUtils.normalizeAnchorUsd(amount / rate);
           debugPrint(
-            '\u001b[41m[GLOBAL_FORM][CALC] amount=$amount, currency=USD, anchorUsdValue=$anchorUsdValue\u001b[0m',
+            '\u001b[41m[GLOBAL_FORM][CALC] amount=$amountStr, currency=$_currencyCode, rate=$rate, anchorUsdValue=$anchorUsdValue\u001b[0m',
           );
         } else {
-          anchorUsdValue = null;
-          debugPrint(
-            '\u001b[41m[GLOBAL_FORM][CALC][WARN] No rate for currency=$_currencyCode, anchorUsdValue=null\u001b[0m',
-          );
+          setState(() {
+            _error = 'No existe una tasa válida para la moneda seleccionada.';
+            _loading = false;
+          });
+          logError('No rate for currency=$_currencyCode');
+          return;
         }
       } else {
         anchorUsdValue = null;
@@ -207,12 +310,13 @@ class _GlobalTransactionFormState extends State<_GlobalTransactionForm> {
         _selectedDate.day,
       );
 
+      /// Guardar el monto como string formateado en la transacción
       final transaction = Transaction(
         id: localId, // id local único
         clientId: _selectedClient!.id,
         userId: widget.userId,
         type: _type!,
-        amount: amount,
+        amount: amountDouble, // <--- double para cálculos y backend
         description: _descriptionController.text,
         date: normalizedDate,
         createdAt: now,
@@ -293,15 +397,14 @@ class _GlobalTransactionFormState extends State<_GlobalTransactionForm> {
     final colorScheme = Theme.of(context).colorScheme;
     final symbol = "";
     final currencyProvider = Provider.of<CurrencyProvider>(context);
-    // final availableCurrencies = currencyProvider.availableCurrencies; // Eliminada variable no usada
-    final allowedCurrencies = CurrencyProvider.allowedCurrencies;
-    // Eliminada variable no usada: nonUsdCurrencies
-    final codeUC = _currencyCode?.toUpperCase();
+    final availableCurrencies = currencyProvider.availableCurrencies;
+    final rate = _currencyCode != null
+        ? currencyProvider.getRateFor(_currencyCode!)
+        : null;
     final rateMissing =
-        codeUC != null &&
-        codeUC != 'USD' &&
-        (currencyProvider.exchangeRates[codeUC] == null ||
-            currencyProvider.exchangeRates[codeUC] == 0);
+        _currencyCode != null &&
+        _currencyCode!.toUpperCase() != 'USD' &&
+        (rate == null || rate == 0);
     _rateFieldVisible = rateMissing;
     final rateValid =
         double.tryParse(_rateController.text.replaceAll(',', '.')) != null &&
@@ -423,36 +526,8 @@ class _GlobalTransactionFormState extends State<_GlobalTransactionForm> {
             },
           ),
         ),
-        Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Icon(
-              _type == 'debt'
-                  ? Icons.trending_down
-                  : _type == 'payment'
-                  ? Icons.trending_up
-                  : Icons.swap_horiz,
-              color: _type == 'debt'
-                  ? Colors.red
-                  : _type == 'payment'
-                  ? Colors.green
-                  : Colors.grey,
-              size: 32,
-            ),
-            const SizedBox(width: 8),
-            Text(
-              _type == 'debt'
-                  ? 'Registrar Deuda'
-                  : _type == 'payment'
-                  ? 'Registrar Abono'
-                  : 'Selecciona tipo de movimiento',
-              style: Theme.of(
-                context,
-              ).textTheme.titleMedium?.copyWith(fontWeight: FontWeight.bold),
-            ),
-          ],
-        ),
-        const SizedBox(height: 16),
+        // Icono + texto (estado/tipo) removido por solicitud
+        const SizedBox(height: 10),
         Center(
           child: AnimatedContainer(
             duration: const Duration(milliseconds: 200),
@@ -485,54 +560,52 @@ class _GlobalTransactionFormState extends State<_GlobalTransactionForm> {
           ),
         ),
         const SizedBox(height: 12),
-        Row(
-          children: [
-            Expanded(
-              child: TextField(
-                controller: _amountController,
-                decoration: InputDecoration(
-                  labelText: 'Monto',
-                  border: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(8),
-                  ),
-                  prefixIcon: Padding(
-                    padding: const EdgeInsets.only(left: 8, right: 4),
-                    child: Text(
-                      symbol,
-                      style: const TextStyle(
-                        fontSize: 18,
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                  ),
-                  prefixIconConstraints: const BoxConstraints(
-                    minWidth: 0,
-                    minHeight: 0,
-                  ),
-                  isDense: true,
+        // Campo monto
+        TextField(
+          controller: _amountController,
+          decoration: InputDecoration(
+            labelText: 'Monto',
+            border: OutlineInputBorder(borderRadius: BorderRadius.circular(8)),
+            prefixIcon: Padding(
+              padding: const EdgeInsets.only(left: 8, right: 4),
+              child: Text(
+                symbol,
+                style: const TextStyle(
+                  fontSize: 18,
+                  fontWeight: FontWeight.bold,
                 ),
-                keyboardType: TextInputType.numberWithOptions(decimal: true),
-                textInputAction: TextInputAction.next,
               ),
             ),
-            const SizedBox(width: 8),
-            SizedBox(
-              width: 110,
+            prefixIconConstraints: const BoxConstraints(
+              minWidth: 0,
+              minHeight: 0,
+            ),
+            isDense: true,
+          ),
+          keyboardType: TextInputType.numberWithOptions(decimal: true),
+          textInputAction: TextInputAction.next,
+          focusNode: _amountFocusNode,
+          inputFormatters: [_thousandsFormatter],
+        ),
+        const SizedBox(height: 7),
+        // Fila de selector de moneda + botón agregar
+        Row(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Expanded(
               child: DropdownButtonFormField<String>(
                 value: _currencyCode,
                 decoration: const InputDecoration(
-                  labelText: 'Moneda',
+                  labelText: 'Tipo de Moneda',
                   border: OutlineInputBorder(),
                   isDense: true,
                 ),
-                items: [
-                  ...[
-                    'USD',
-                    ...allowedCurrencies.where((code) => code != 'USD'),
-                  ].map(
-                    (code) => DropdownMenuItem(value: code, child: Text(code)),
-                  ),
-                ],
+                items: availableCurrencies
+                    .map(
+                      (code) =>
+                          DropdownMenuItem(value: code, child: Text(code)),
+                    )
+                    .toList(),
                 onChanged: (code) {
                   setState(() {
                     _currencyCode = code;
@@ -543,6 +616,91 @@ class _GlobalTransactionFormState extends State<_GlobalTransactionForm> {
                 menuMaxHeight: 180,
               ),
             ),
+            const SizedBox(width: 6),
+            SizedBox(
+              height: 48,
+              child: IconButton(
+                icon: const Icon(
+                  Icons.add_circle_outline,
+                  color: Colors.indigo,
+                  size: 24,
+                ),
+                tooltip: 'Agregar moneda',
+                onPressed: () async {
+                  String? newCode = await showDialog<String>(
+                    context: context,
+                    builder: (ctx) {
+                      final controller = TextEditingController();
+                      return AlertDialog(
+                        title: const Text('Agregar moneda'),
+                        content: TextField(
+                          controller: controller,
+                          decoration: const InputDecoration(
+                            labelText: 'Código (ej: EUR)',
+                            border: OutlineInputBorder(),
+                            isDense: true,
+                          ),
+                          textCapitalization: TextCapitalization.sentences,
+                          maxLength: 11,
+                        ),
+                        actions: [
+                          TextButton(
+                            onPressed: () => Navigator.of(ctx).pop(),
+                            child: const Text('Cancelar'),
+                          ),
+                          ElevatedButton(
+                            onPressed: () {
+                              String code = controller.text.trim();
+                              if (code.isEmpty) {
+                                Navigator.of(ctx).pop();
+                                return;
+                              }
+                              // Solo primera letra mayúscula, resto minúscula
+                              code =
+                                  code.substring(0, 1).toUpperCase() +
+                                  (code.length > 1
+                                      ? code.substring(1).toLowerCase()
+                                      : '');
+                              if (code == 'USD' ||
+                                  availableCurrencies.contains(code)) {
+                                Navigator.of(ctx).pop();
+                                return;
+                              }
+                              Navigator.of(ctx).pop(code);
+                            },
+                            child: const Text('Agregar'),
+                          ),
+                        ],
+                      );
+                    },
+                  );
+                  if (newCode != null &&
+                      newCode.isNotEmpty &&
+                      newCode != 'USD' &&
+                      !availableCurrencies.contains(newCode)) {
+                    // Agregar la moneda al provider
+                    final currencyProvider = Provider.of<CurrencyProvider>(
+                      context,
+                      listen: false,
+                    );
+                    currencyProvider.addManualCurrency(newCode);
+
+                    // Buscar la versión realmente insertada (puede cambiar el casing)
+                    String selectedValue = newCode;
+                    for (final c in currencyProvider.availableCurrencies) {
+                      if (c.toLowerCase() == newCode.toLowerCase()) {
+                        selectedValue = c;
+                        break;
+                      }
+                    }
+                    setState(() {
+                      _currencyCode = selectedValue;
+                      _rateController.text = '';
+                    });
+                  }
+                },
+              ),
+            ),
           ],
         ),
         if (_rateFieldVisible)
@@ -551,9 +709,7 @@ class _GlobalTransactionFormState extends State<_GlobalTransactionForm> {
             child: TextField(
               controller: _rateController,
               decoration: InputDecoration(
-                labelText:
-                    'Tasa  {_currencyCode?.toUpperCase() ?? '
-                    '} a USD',
+                labelText: 'Tasa ${_currencyCode?.toUpperCase() ?? ''} a USD',
                 border: OutlineInputBorder(),
                 isDense: true,
                 prefixIcon: Icon(Icons.attach_money_rounded),
@@ -586,17 +742,60 @@ class _GlobalTransactionFormState extends State<_GlobalTransactionForm> {
           ),
         ),
         const SizedBox(height: 12),
-        TextField(
-          controller: _descriptionController,
-          maxLength: 30,
-          decoration: InputDecoration(
-            labelText: 'Descripción',
-            border: OutlineInputBorder(borderRadius: BorderRadius.circular(8)),
-            prefixIcon: Icon(Icons.description),
-            isDense: true,
-            counterText: '',
-          ),
-          maxLines: 2,
+        // Campo descripción con contador visual igual a transaction_form
+        Stack(
+          children: [
+            TextField(
+              controller: _descriptionController,
+              maxLength: 30,
+              decoration: InputDecoration(
+                labelText: 'Descripción',
+                border: OutlineInputBorder(
+                  borderRadius: BorderRadius.circular(8),
+                ),
+                prefixIcon: Icon(Icons.description),
+                isDense: true,
+                counterText: '', // Oculta el contador por defecto
+              ),
+              maxLines: 2,
+              onChanged: (_) => setState(() {}),
+            ),
+            Positioned(
+              right: 12,
+              bottom: 6,
+              child: Container(
+                padding: const EdgeInsets.symmetric(horizontal: 7, vertical: 2),
+                decoration: BoxDecoration(
+                  color: Colors.white,
+                  borderRadius: BorderRadius.circular(8),
+                  border: Border.all(color: Colors.grey.shade300, width: 1),
+                ),
+                child: Builder(
+                  builder: (context) {
+                    final currentLength = _descriptionController.text.length;
+                    final maxLength = 30;
+                    final remaining = maxLength - currentLength;
+                    Color color;
+                    if (remaining <= 5) {
+                      color = Colors.red;
+                    } else if (remaining <= 10) {
+                      color = Colors.orange;
+                    } else {
+                      color = Colors.grey;
+                    }
+                    return Text(
+                      '$currentLength/$maxLength',
+                      style: TextStyle(
+                        fontSize: 11,
+                        color: color,
+                        fontWeight: FontWeight.w500,
+                      ),
+                    );
+                  },
+                ),
+              ),
+            ),
+          ],
         ),
         if (_error != null)
           Padding(
