@@ -48,183 +48,219 @@ class TransactionCard extends StatelessWidget {
         isOffline: isOffline,
       );
     }
-    return Container(
-      decoration: BoxDecoration(
-        color: const Color.fromARGB(255, 255, 255, 255),
-        borderRadius: BorderRadius.circular(16),
-        boxShadow: [
-          BoxShadow(
-            color: const Color.fromARGB(
-              255,
-              11,
-              11,
-              11,
-            ).withAlpha((0.25 * 255).toInt()),
-            blurRadius: 4,
-            offset: const Offset(0, 2),
-          ),
-        ],
-      ),
-      child: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 10),
-        child: Row(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            // --- ICONO IZQUIERDA ---
-            CircleAvatar(
-              backgroundColor: t.type == 'debt'
-                  ? const Color(0xFFFFE5E5)
-                  : const Color(0xFFE5FFE8),
-              radius: 16,
-              child: Icon(
-                t.type == 'debt' ? Icons.arrow_downward : Icons.arrow_upward,
-                color: t.type == 'debt' ? Colors.red : Colors.green,
-                size: 16,
+    return Padding(
+      padding: const EdgeInsets.symmetric(
+        horizontal: 2,
+      ), // Ajusta el valor a tu gusto
+      child: ConstrainedBox(
+        constraints: BoxConstraints(maxWidth: 300),
+        child: Container(
+          decoration: BoxDecoration(
+            color: const Color.fromARGB(255, 255, 255, 255),
+            borderRadius: BorderRadius.circular(16),
+            boxShadow: [
+              BoxShadow(
+                color: const Color.fromARGB(
+                  255,
+                  11,
+                  11,
+                  11,
+                ).withAlpha((0.25 * 255).toInt()),
+                blurRadius: 4,
+                offset: const Offset(0, 2),
               ),
-            ),
-            const SizedBox(width: 10),
-
-            // --- CONTENIDO PRINCIPAL ---
-            Expanded(
-              child: Column(
-                children: [
-                  // --- ROW 1: NOMBRE CLIENTE ---
-                  SizedBox(
-                    width: double.infinity,
-                    child: Text(
-                      client.name,
-                      style: const TextStyle(
-                        fontWeight: FontWeight.bold,
-                        fontSize: 16,
-                      ),
-                      textAlign: TextAlign.left,
-                      maxLines: 1,
-                      overflow: TextOverflow.ellipsis,
-                    ),
+            ],
+          ),
+          child: Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 10),
+            child: Row(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                // --- ICONO IZQUIERDA ---
+                CircleAvatar(
+                  backgroundColor: t.type == 'debt'
+                      ? const Color(0xFFFFE5E5)
+                      : const Color(0xFFE5FFE8),
+                  radius: 16,
+                  child: Icon(
+                    t.type == 'debt'
+                        ? Icons.arrow_downward
+                        : Icons.arrow_upward,
+                    color: t.type == 'debt' ? Colors.red : Colors.green,
+                    size: 16,
                   ),
-                  const SizedBox(height: 4),
+                ),
+                const SizedBox(width: 10),
 
-                  // --- ROW 2: DESCRIPCIÓN Y FECHA ---
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    crossAxisAlignment: CrossAxisAlignment.start,
+                // --- CONTENIDO PRINCIPAL ---
+                Expanded(
+                  child: Column(
                     children: [
-                      Expanded(
+                      // --- ROW 1: NOMBRE CLIENTE ---
+                      SizedBox(
+                        width: double.infinity,
                         child: Text(
-                          t.description != null &&
-                                  t.description is String &&
-                                  t.description.isNotEmpty
-                              ? t.description[0].toUpperCase() +
-                                    t.description.substring(1)
-                              : 'Sin descripción',
+                          client.name,
                           style: const TextStyle(
-                            fontSize: 13.5,
-                            color: Colors.black54,
+                            fontWeight: FontWeight.bold,
+                            fontSize: 16,
                           ),
+                          textAlign: TextAlign.left,
+                          maxLines: 1,
+                          overflow: TextOverflow.ellipsis,
                         ),
                       ),
-                      const SizedBox(width: 8),
-                      Text(
-                        '${t.date.year}-${t.date.month.toString().padLeft(2, '0')}-${t.date.day.toString().padLeft(2, '0')}',
-                        style: const TextStyle(
-                          fontSize: 12.5,
-                          color: Colors.black45,
-                        ),
-                      ),
-                    ],
-                  ),
-                  const SizedBox(height: 6),
+                      const SizedBox(height: 4),
 
-                  // --- ROW 3: MONTO / BALANCE ---
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.end,
-                    children: [
-                      Column(
-                        crossAxisAlignment: CrossAxisAlignment.end,
+                      // --- ROW 2: DESCRIPCIÓN Y FECHA ---
+                      Row(
+                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                        crossAxisAlignment: CrossAxisAlignment.start,
                         children: [
-                          Text(
-                            () {
-                              final usdValue =
-                                  t.anchorUsdValue ?? t.amount ?? 0.0;
-                              if (selectedCurrency == 'USD') {
-                                return 'USD ${CurrencyUtils.formatNumber(usdValue)}';
-                              } else {
-                                final rate =
-                                    exchangeRates[selectedCurrency] ?? 1.0;
-                                final convertedValue = usdValue * rate;
-                                return '${CurrencyUtils.formatNumber(convertedValue)} $selectedCurrency';
-                              }
-                            }(),
-                            style: TextStyle(
-                              color: t.type == 'payment'
-                                  ? Colors.green
-                                  : Colors.red,
-                              fontWeight: FontWeight.bold,
-                              fontSize: 18,
-                            ),
-                          ),
-                          if (selectedCurrency != 'USD')
-                            Padding(
-                              padding: const EdgeInsets.only(top: 2),
-                              child: Text(
-                                'USD ${CurrencyUtils.formatNumber(t.anchorUsdValue ?? t.amount ?? 0.0)}',
-                                style: const TextStyle(
-                                  fontSize: 13,
-                                  color: Colors.black54,
-                                  fontWeight: FontWeight.w500,
-                                ),
+                          Expanded(
+                            child: Text(
+                              t.description != null &&
+                                      t.description is String &&
+                                      t.description.isNotEmpty
+                                  ? t.description[0].toUpperCase() +
+                                        t.description.substring(1)
+                                  : 'Sin descripción',
+                              style: const TextStyle(
+                                fontSize: 13.5,
+                                color: Colors.black54,
                               ),
                             ),
+                          ),
+                          const SizedBox(width: 8),
+                          Text(
+                            '${t.date.year}-${t.date.month.toString().padLeft(2, '0')}-${t.date.day.toString().padLeft(2, '0')}',
+                            style: const TextStyle(
+                              fontSize: 12.5,
+                              color: Colors.black45,
+                            ),
+                          ),
                         ],
                       ),
-                    ],
-                  ),
+                      const SizedBox(height: 6),
 
-                  // --- MENSAJE DE SINCRONIZACIÓN (ABAJO) ---
-                  if (syncMsg != null)
-                    Padding(
-                      padding: const EdgeInsets.only(top: 8, bottom: 1),
-                      child: Row(
+                      // --- ROW 3: MONTO / BALANCE ---
+                      Row(
+                        mainAxisAlignment: MainAxisAlignment.end,
                         children: [
-                          const Spacer(),
-                          Container(
-                            padding: const EdgeInsets.symmetric(
-                              horizontal: 7,
-                              vertical: 2,
-                            ),
-                            decoration: BoxDecoration(
-                              color: syncMsg.color.withAlpha(
-                                (0.09 * 255).toInt(),
-                              ),
-                              borderRadius: BorderRadius.circular(10),
-                            ),
-                            child: Row(
-                              mainAxisSize: MainAxisSize.min,
-                              children: [
-                                Icon(
-                                  syncMsg.icon,
-                                  size: 12,
-                                  color: syncMsg.color,
+                          Column(
+                            crossAxisAlignment: CrossAxisAlignment.end,
+                            children: [
+                              // --- PATCH: Visualización correcta de montos ---
+                              Text(
+                                () {
+                                  final txCurrency = t.currencyCode ?? 'USD';
+                                  final rateForSelected =
+                                      exchangeRates[selectedCurrency] ?? 1.0;
+                                  double anchorUsd;
+                                  if (t.anchorUsdValue != null) {
+                                    anchorUsd = t.anchorUsdValue as double;
+                                  } else if (t.amount != null &&
+                                      t.originalRate != null &&
+                                      t.originalRate > 0) {
+                                    anchorUsd = t.amount / t.originalRate;
+                                  } else if (t.amount != null) {
+                                    anchorUsd = t.amount;
+                                  } else {
+                                    anchorUsd = 0.0;
+                                  }
+
+                                  if (selectedCurrency == txCurrency) {
+                                    debugPrint(
+                                      '[CARD][MONTO] id: ${t.id}, tipo: ${t.type}, montoOriginal: ${t.amount}, moneda: $selectedCurrency',
+                                    );
+                                    return '${CurrencyUtils.formatNumber(t.amount ?? 0.0)} $selectedCurrency';
+                                  } else if (selectedCurrency == 'USD') {
+                                    debugPrint(
+                                      '[CARD][MONTO] id: ${t.id}, tipo: ${t.type}, montoUSD: $anchorUsd, moneda: USD',
+                                    );
+                                    return 'USD ${CurrencyUtils.formatNumber(anchorUsd)}';
+                                  } else {
+                                    final converted =
+                                        anchorUsd * rateForSelected;
+                                    debugPrint(
+                                      '[CARD][MONTO] id: ${t.id}, tipo: ${t.type}, montoUSD: $anchorUsd, tasa: $rateForSelected, convertido: $converted, moneda: $selectedCurrency',
+                                    );
+                                    return '${CurrencyUtils.formatNumber(converted)} $selectedCurrency';
+                                  }
+                                }(),
+                                style: TextStyle(
+                                  color: t.type == 'payment'
+                                      ? Colors.green
+                                      : Colors.red,
+                                  fontWeight: FontWeight.bold,
+                                  fontSize: 18,
                                 ),
-                                const SizedBox(width: 2),
-                                Text(
-                                  syncMsg.message,
-                                  style: TextStyle(
-                                    fontSize: 9,
-                                    color: syncMsg.color,
-                                    fontWeight: FontWeight.w600,
+                              ),
+                              // Si la moneda seleccionada no es USD, muestra el monto en USD debajo
+                              if (selectedCurrency != 'USD')
+                                Padding(
+                                  padding: const EdgeInsets.only(top: 2),
+                                  child: Text(
+                                    'USD ${CurrencyUtils.formatNumber((t.anchorUsdValue ?? (t.amount ?? 0.0)) as double)}',
+                                    style: const TextStyle(
+                                      fontSize: 13,
+                                      color: Colors.black54,
+                                      fontWeight: FontWeight.w500,
+                                    ),
                                   ),
                                 ),
-                              ],
-                            ),
+                            ],
                           ),
                         ],
                       ),
-                    ),
-                ],
-              ),
+
+                      // --- MENSAJE DE SINCRONIZACIÓN (ABAJO) ---
+                      if (syncMsg != null)
+                        Padding(
+                          padding: const EdgeInsets.only(top: 8, bottom: 1),
+                          child: Row(
+                            children: [
+                              const Spacer(),
+                              Container(
+                                padding: const EdgeInsets.symmetric(
+                                  horizontal: 7,
+                                  vertical: 2,
+                                ),
+                                decoration: BoxDecoration(
+                                  color: syncMsg.color.withAlpha(
+                                    (0.09 * 255).toInt(),
+                                  ),
+                                  borderRadius: BorderRadius.circular(10),
+                                ),
+                                child: Row(
+                                  mainAxisSize: MainAxisSize.min,
+                                  children: [
+                                    Icon(
+                                      syncMsg.icon,
+                                      size: 12,
+                                      color: syncMsg.color,
+                                    ),
+                                    const SizedBox(width: 2),
+                                    Text(
+                                      syncMsg.message,
+                                      style: TextStyle(
+                                        fontSize: 9,
+                                        color: syncMsg.color,
+                                        fontWeight: FontWeight.w600,
+                                      ),
+                                    ),
+                                  ],
+                                ),
+                              ),
+                            ],
+                          ),
+                        ),
+                    ],
+                  ),
+                ),
+              ],
             ),
-          ],
+          ),
         ),
       ),
     );
